HBASE-19239 Fix findbugs and error-prone issues
Fixes for hbase-common

+++ hbase-common/src/main/java/org/apache/hadoop/hbase/CellUtil.java
@@ -633,7 +633,7 @@ public final class CellUtil {
     if (cell instanceof KeyValue) {
       return ((KeyValue)cell).heapSizeWithoutTags();
     }
-    return getSumOfCellKeyElementLengths(cell) + cell.getValueLength();
+    return (long) getSumOfCellKeyElementLengths(cell) + cell.getValueLength();
   }
 
   /********************* tags *************************************/
