+++ hbase-common/src/main/java/org/apache/hadoop/hbase/util/ClassSize.java
@@ -344,7 +344,7 @@ public class ClassSize {
    * @return the size estimate, in bytes
    */
   private static long estimateBaseFromCoefficients(int [] coeff, boolean debug) {
-    long prealign_size = OBJECT + coeff[0] + coeff[2] * REFERENCE;
+    long prealign_size = (long) OBJECT + coeff[0] + coeff[2] * REFERENCE;
 
     // Round up to a multiple of 8
     long size = align(prealign_size) + align(coeff[1] * ARRAY);
